[requirement] DEVT-691 GPU support mechanical lidar optical centre correction

+++ libhesai/UdpParserGpu/include/udp2_6_parser_gpu.h
@@ -55,8 +55,8 @@ class Udp2_6ParserGpu: public GeneralParserGpu<T_Point>{
  private:
   float* channel_azimuths_cu_;
   float* channel_elevations_cu_;
-  int16_t* channel_azimuths_adjust_cu_;
-  int16_t* channel_elevations_adjust_cu_;
+  float* channel_azimuths_adjust_cu_;
+  float* channel_elevations_adjust_cu_;
   float* raw_azimuths_cu_;
   float* raw_elevations_cu_;
   uint16_t* raw_distances_cu_;
